Move DBEdge into model package
