+++ src/main/java/com/github/bachelorpraktikum/dbvisualization/database/model/Attribute.java
@@ -1,7 +1,6 @@
 package com.github.bachelorpraktikum.dbvisualization.database.model;
 
 import com.github.bachelorpraktikum.dbvisualization.config.ConfigKey;
-import com.github.bachelorpraktikum.dbvisualization.database.DBEdge;
 import java.sql.ResultSet;
 import java.sql.SQLException;
 import java.util.Collections;
