+++ src/main/java/com/github/bachelorpraktikum/dbvisualization/database/model/Vertex.java
@@ -1,6 +1,5 @@
 package com.github.bachelorpraktikum.dbvisualization.database.model;
 
-import com.github.bachelorpraktikum.dbvisualization.database.DBEdge;
 import com.github.bachelorpraktikum.dbvisualization.model.Coordinates;
 import java.sql.ResultSet;
 import java.sql.SQLException;
